bench(holon-openai): holon-0822-runtime-work-item-test-split
