+++ src/runtime/tests/mod.rs
@@ -2,5 +2,8 @@ mod agent_and_tools;
 mod runtime_state;
 mod support;
 mod turns;
-mod work_items;
 mod workspace;
+mod work_items;
+mod task_recovery;
+mod timers;
+mod wake_hints;
